requirements.txt
fastapi
uvicorn
python-multipart
dnspython
